backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
google-generativeai>=0.3.0
python-dotenv==1.0.0
python-multipart>=0.0.9
pydantic>=2.0.0
